refactor: extract error and FPI utility functions
