+++ src/utils/index.ts
@@ -1,6 +1,12 @@
 import { ILogger } from '../types';
+import { createErrorUtils } from './error';
+import { createFpiUtils } from './fpi';
 import { createHttpUtils } from './http';
 
 export function createUtils(logger: ILogger) {
-  return { ...createHttpUtils(logger) };
+  return { 
+    ...createHttpUtils(logger),
+    ...createErrorUtils(logger),
+    ...createFpiUtils(logger),
+  };
 }
